requirements.txt
ffmpeg~=1.4
ffmpeg-python~=0.2.0
Flask~=3.1.2
Flask-Login~=0.6.3
Flask-Migrate~=4.1.0
SQLAlchemy~=2.0.40
Werkzeug~=3.1.3
Flask-Bcrypt~=1.0.1
alembic~=1.16.5
Flask-Moment~=1.0.6
Flask-SQLAlchemy~=3.1.1
psycopg2-binary==2.9.10
gunicorn
Markdown~=3.8
filetype_py~=1.0
flask_mail~=0.10.0
